@@ -7,8 +7,8 @@
     "@testing-library/react": "^12.0.0",
     "@testing-library/user-event": "^13.2.1",
     "moment": "^2.29.4",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-paginate": "^8.1.3",
     "react-scripts": "4.0.3",
     "react-select": "^5.5.4",
