@@ -14,9 +14,14 @@ import './styles/index.css';
 // Include application component.
 import App from './components/App';
 
-ReactDOM.render(
+// React 18 render
+import {createRoot} from 'react-dom/client';
+
+const container = document.getElementById('root')
+const root = createRoot(container)
+root.render(
   <React.StrictMode>
     <App />
   </React.StrictMode>,
-  document.getElementById('root')
-);
+)
+
